microservices/MetadataStorage/tests/requirements.txt
pytest==8.2.2
pytest-cov==5.0.0
pytest-asyncio~=0.23.7
pytest-redis==3.1.2
httpx==0.27.0
factory-boy==3.3
freezegun==1.5.1